+++ examples/SIR.hs
@@ -4,258 +4,254 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators, TypeApplications #-}
 {-# LANGUAGE OverloadedLabels #-}
+{-# LANGUAGE PolyKinds #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE TypeFamilyDependencies #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
 {-# HLINT ignore "Redundant return" #-}
+
+{- | This demonstrates:
+      - The [SIR](https://en.wikipedia.org/wiki/Compartmental_models_in_epidemiology) model for modelling
+        the transition between Susceptible (S), Infected (I), and Recovered (R) individuals over a number of days
+        during an epidemic. We model this as a Hidden Markov Model, where the latent states are the true values
+        of S, I, and R, and the observations are the reported number of infections (𝜉).
+      - Extending the SIR to the SIRS model where recovered individuals (R) can become susceptible (S) again.
+      - Extending the SIRS to the SIRSV model where susceptible individuals (S) can become vaccinated (V).
+
+    For convenience, this makes use of the 'Data.Extensible' library for extensible records, and the 'Control.Lens'
+    library to record accessors. If the lens notation is unfamiliar, the code below can be corresponded to a less modular
+    version the file [SIRNonModular](examples/SIRNonModular.hs).
+ -}
+
 module SIR where
 
-import Prog
-import Effects.Writer
-import Model
-import Inference.SIM as SIM
-import Inference.MH as MH
-import Sampler
-import Env
-import Control.Monad
-import qualified Control.Monad.Bayes.Class as Bayes
-import qualified Control.Monad.Bayes.Weighted as Bayes
-import qualified Control.Monad.Bayes.Traced as Bayes
-import qualified Control.Monad.Bayes.Sampler as Bayes
-import Trace
-import Inference.MB as MB
+import Data.Extensible ( mkField, Record, Lookup, type (>:), (@=), (<:), emptyRecord, Assoc ((:>)) )
+import Prog ( Member )
+import Control.Lens ( (&), (^.), (.~) )
+import Effects.Writer ( Writer, tellM, handleWriterM )
+import Model ( Model, beta, binomial', gamma, poisson )
+import Control.Monad ( (>=>) )
+import Env ( Observables, Observable, Assign ((:=)), (<:>), enil, (<#>), vnil, get)
+import HMM ( ObsModel, TransModel, hmmGen )
+import GHC.TypeLits ( Symbol )
+import Data.Kind (Constraint)
+import Sampler ( Sampler )
+import Inference.SIM as SIM ( simulate )
+import Inference.MH as MH ( mh )
 
-import HMM
-import Data.Extensible (Associated)
+-- | A type family for conveniently specifying multiple @Record@ fields of the same type
+type family Lookups env (ks :: [Symbol]) a :: Constraint where
+  Lookups env (x ': xs) a = (Lookup env x a, Lookups env xs a)
+  Lookups env '[] a = ()
 
--- ||| (Section 3.1 + Section 5.5 extension) The SIR model
-data Popl = Popl {
-    s   :: Int, -- ^ Number of people susceptible to infection
-    i   :: Int, -- ^ Number of people currently infected
-    r   :: Int  -- ^ Number of people recovered from infection
-} deriving Show
+-- | HMM latent states (SIRV)
+mkField "s" -- ^ susceptible individuals
+mkField "i" -- ^ infected individuals
+mkField "r" -- ^ recovered individuals
+mkField "v" -- ^ vaccinated individuals
 
+-- | HMM observations (𝜉) i.e. reported infections
 type Reported = Int
 
--- | SIR transition model
-transSI :: TransModel env ts Double Popl
-transSI beta (Popl s i r) = do
-  let pop = s + i + r
-  dN_SI <- binomial' s (1 - exp ((-beta * fromIntegral i) / fromIntegral pop))
-  return $ Popl (s - dN_SI) (i + dN_SI) r
 
-transIR :: TransModel env ts Double Popl
-transIR gamma (Popl s i r)  = do
-  dN_IR <- binomial' i (1 - exp (-gamma))
-  return $ Popl s (i - dN_IR) (r + dN_IR)
+{- | SIR model.
+-}
 
-data TransParamsSIR = TransParamsSIR {
-    betaP  :: Double, -- ^ Mean contact rate between susceptible and infected people
-    gammaP :: Double  -- ^ Mean recovery rate
-}
+-- | SIR transition prior
+transPriorSIR :: Observables env '["β",  "γ"] Double
+  => Model env ts (Double, Double)
+transPriorSIR = do
+  pBeta  <- gamma 2 1 #β
+  pGamma <- gamma 1 (1/8) #γ
+  return (pBeta, pGamma)
 
-transSIR :: Member (Writer [Popl]) es -- || Writer effect from Section 5.5
-  => TransModel env es TransParamsSIR Popl
-transSIR (TransParamsSIR beta gamma) sir = do
-  sir' <- (transSI beta >=> transIR gamma) sir
-  tellM [sir']
-  return sir'
+-- | Transition model from S and I
+transSI :: Lookups popl '["s", "i", "r"] Int => TransModel env ts Double (Record popl)
+transSI  beta popl = do
+  let (s_0, i_0, r_0 ) = (popl ^. s,  popl ^. i,  popl ^. r)
+      pop = s_0 + i_0 + r_0
+  dN_SI <- binomial' s_0 (1 - exp ((-beta * fromIntegral i_0) / fromIntegral pop))
+  return $ popl & s .~ (s_0 - dN_SI)
+                & i .~ (i_0 + dN_SI)
+
+-- | Transition model from I and R
+transIR :: Lookups popl '["i", "r"] Int => TransModel env ts Double (Record popl)
+transIR  gamma popl = do
+  let (i_0, r_0) = (popl ^. i,  popl ^. r)
+  dN_IR <- binomial' i_0 (1 - exp (-gamma))
+  return $ popl & i .~ (i_0 - dN_IR)
+                & r .~ (r_0 + dN_IR)
+
+-- | Transition model from S to I, and I to R
+transSIR :: (Member (Writer [Record popl]) ts, Lookups popl '["s", "i", "r"] Int)
+  => TransModel env ts (Double, Double) (Record popl)
+transSIR (beta, gamma) popl = do
+  popl <- (transSI beta >=> transIR gamma) popl
+  tellM [popl]  -- a user effect for writing each latent SIR state to a stream [Record popl]
+  return popl
 
 -- | SIR observation model
 type ObsParams = Double
 
-obsSIR :: Observable env "𝜉" Int
-  => ObsModel env ts Double Popl Reported
-obsSIR rho (Popl _ i _)  = do
-  i <- poisson (rho * fromIntegral i) #𝜉
-  return i
-
--- | SIR transition prior
-transPriorSIR :: Observables env '["β",  "γ"] Double
-  => Model env ts TransParamsSIR
-transPriorSIR = do
-  pBeta  <- gamma 2 1 #β
-  pGamma <- gamma 1 (1/8) #γ
-  return (TransParamsSIR pBeta pGamma)
-
 -- | SIR observation prior
 obsPriorSIR :: Observables env '["ρ"] Double
-  => Model env ts ObsParams
-obsPriorSIR = do
-  pRho <- beta 2 7 #ρ
-  return pRho
+  => Model env ts Double
+obsPriorSIR = beta 2 7 #ρ
+
+-- | SIR observation model
+obsSIR :: Lookup s "i" Int => Observable env "𝜉" Int
+  => ObsModel env ts Double (Record s) Reported
+obsSIR rho popl  = do
+  let i_0 = popl ^. i
+  poisson (rho * fromIntegral i_0) #𝜉
 
 -- | SIR as HMM
-hmmSIR :: (Member (Writer [Popl]) es, Observable env "𝜉" Int, Observables env '["ρ", "β", "γ"] Double)
-  => Int -> Popl -> Model env es Popl
-hmmSIR = hmmGen transPriorSIR obsPriorSIR transSIR obsSIR
+hmmSIR :: (Lookups popl '["s", "i", "r"] Int
+        , Observable env "𝜉" Int
+        , Observables env '["ρ", "β", "γ"] Double)
+  -- | number of days
+  => Int
+  -- | initial population state
+  -> Record popl
+  -- | (final population state, intermediate population states)
+  -> Model env es (Record popl, [Record popl])
+hmmSIR n = handleWriterM . hmmGen transPriorSIR obsPriorSIR transSIR obsSIR n
 
-hmmSIR' :: (Observables env '["𝜉"] Int , Observables env '[ "β" , "γ" , "ρ"] Double) => Int -> Popl -> Model env es (Popl, [Popl])
-hmmSIR' n = handleWriterM . hmmSIR n
-
-type SIRenv = '["β" := Double, "γ"  := Double, "ρ"  := Double, "𝜉" := Int]
-
--- ||| (Section 3.1, Fig 4a) SIM from SIR model: ([(s, i, r)], [𝜉])
+-- | Simulate from the SIR model
 simSIR :: Int -> Sampler ([(Int, Int, Int)], [Reported])
 simSIR n_days = do
-  let sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
-      sir_0      = Popl {s = 762, i = 1, r = 0}
-  ((_, sir_trace), sim_env_out) <- SIM.simulate (hmmSIR' n_days sir_0) sim_env_in
+  -- Specify model input of 762 susceptible and 1 infected
+  let sir_0      = #s @= 762 <: #i @= 1 <: #r @= 0 <: emptyRecord
+  -- Specify model environment
+      sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
+  -- Simulate an epidemic over 100 days
+  ((_, sir_trace), sim_env_out) <- SIM.simulate (hmmSIR n_days sir_0) sim_env_in
+  -- Get the observed infections over 100 days
   let 𝜉s :: [Reported] = get #𝜉 sim_env_out
-      sirs = map (\(Popl s i recov) -> (s, i, recov)) sir_trace
+  -- Get the true SIR values over 100 days
+      sirs = map (\sir -> (sir ^. s, sir ^. i, sir ^. r)) sir_trace
   return (sirs, 𝜉s)
 
--- ||| (Section 3.3, Fig 5) Infer from SIR model: ([ρ], [β])
+-- | MH inference from SIR model: ([ρ], [β])
 mhSIR :: Int -> Int -> Sampler ([Double], [Double])
 mhSIR n_mhsteps n_days = do
+  -- Simulate some observed infections
   𝜉s <- snd <$> simSIR n_days
-  let mh_env_in = #β := [] <:> #γ := [0.0085] <:> #ρ := [] <:> #𝜉 := 𝜉s <:> enil
-      sir_0           = Popl {s = 762, i = 1, r = 0}
-  mhTrace <- MH.mh n_mhsteps (hmmSIR' n_days sir_0) mh_env_in (#β <#> #ρ <#> vnil)
+  -- Specify model input of 762 susceptible and 1 infected
+  let sir_0     = #s @= 762 <: #i @= 1 <: #r @= 0 <: emptyRecord
+  -- Specify model environment
+      mh_env_in = #β := [] <:> #γ := [0.0085] <:> #ρ := [] <:> #𝜉 := 𝜉s <:> enil
+  -- Run MH inference over 50000 iterations
+  mhTrace <- MH.mh n_mhsteps (hmmSIR n_days sir_0) mh_env_in (#β <#> #ρ <#> vnil)
+  -- Get the sampled values for model parameters ρ and β
   let ρs = concatMap (get #ρ) mhTrace
       βs = concatMap (get #β) mhTrace
   return (ρs, βs)
 
--- ||| SIR model using monad-bayes inference
-mbayesSIR ::
-   (FromSTrace env, Bayes.MonadInfer m
-  , Observables env '["𝜉"] Int , Observables env '[ "β" , "γ" , "ρ"] Double)
-  => Int -> Popl -> Env env -> m ((Popl, [Popl]), Env env)
-mbayesSIR n popl = toMBayes (hmmSIR' n popl)
 
-simSIRMB :: Int -> IO ([(Int, Int, Int)], [Reported])
-simSIRMB n_days = do
-  let sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
-      sir_0      = Popl {s = 762, i = 1, r = 0}
-  ((_, sir_trace), sim_env_out) <- Bayes.sampleIO $ Bayes.prior (mbayesSIR n_days sir_0 sim_env_in)
-  let 𝜉s :: [Reported] = get #𝜉 sim_env_out
-      sirs = map (\(Popl s i recov) -> (s, i, recov)) sir_trace
-  pure (sirs, 𝜉s)
+{- | SIRS model.
+-}
 
-mhSIRMB :: Int -> IO ([Double], [Double]) -- [(Popl, Env SIRenv)]
-mhSIRMB n_days = do
-  𝜉s <- snd <$> simSIRMB n_days
-  let sir_0      = Popl {s = 762, i = 1, r = 0}
-      env = #β := [] <:> #γ := [0.0085] <:> #ρ := [] <:> #𝜉 := 𝜉s <:> enil
-  (_, env) <- unzip <$> Bayes.sampleIO (Bayes.prior $ Bayes.mh 100 (mbayesSIR 100 sir_0 env))
-  let ρs = concatMap (get #ρ) env
-      βs = concatMap (get #β) env
-  pure (ρs, βs)
-
--- ||| (Section 3.2) Modular Extensions to the SIR Model
-{- Note that the implementations below aren't as modular as we would like, due to having to redefine the data types Popl and TransParams when adding new variables to the SIR model. The file "src/Examples/SIRModular.hs" shows how one could take steps to resolve this by using extensible records. -}
-
--- || (Section 3.2)  SIRS (resusceptible) model
-data TransParamsSIRS = TransParamsSIRS {
-    betaP_SIRS  :: Double, -- ^ Mean contact rate between susceptible and infected people
-    gammaP_SIRS :: Double, -- ^ Mean recovery rate
-    etaP_SIRS   :: Double  -- ^ Rate of resusceptible
-}
-
--- | SIRS transition model
-transRS :: Double -> Popl -> Model env ts Popl
-transRS eta (Popl s i r) = do
-  dN_RS <- binomial' r (1 - exp (-eta))
-  return $ Popl (s + dN_RS) i (r - dN_RS)
-
-transSIRS :: Member (Writer [Popl]) es
-  => TransModel env es TransParamsSIRS Popl
-transSIRS (TransParamsSIRS beta gamma eta) sir = do
-  sir' <- (transSI beta >=> transIR gamma >=> transRS eta) sir
-  tellM [sir']
-  return sir'
-
--- | SIR transition prior
+-- | SIRS transition prior
 transPriorSIRS :: Observables env '["β", "η", "γ"] Double
-  => Model env ts TransParamsSIRS
+  => Model env ts (Double, Double, Double)
 transPriorSIRS = do
-  TransParamsSIR pBeta pGamma  <- transPriorSIR
+  (pBeta, pGamma)  <- transPriorSIR
   pEta <- gamma 1 (1/8) #η
-  return (TransParamsSIRS pBeta pGamma pEta)
+  return (pBeta, pGamma, pEta)
+
+-- | Transition model from S to R
+transRS :: Lookups popl '["s", "r"] Int => TransModel env ts Double (Record popl)
+transRS eta popl = do
+  let (r_0, s_0) = (popl ^. r,  popl ^. s)
+  dN_RS <- binomial' r_0 (1 - exp (-eta))
+  return $ popl & r .~ (r_0 - dN_RS)
+                & s .~ (s_0 + dN_RS)
+
+-- | Transition model from S to I, I to R, and R to S
+transSIRS :: Lookups popl '["s", "i", "r"] Int => TransModel env es (Double, Double, Double) (Record popl)
+transSIRS (beta, gamma, eta) = transSI beta >=> transIR gamma >=> transRS eta
 
 -- | SIRS as HMM
-hmmSIRS :: (Observables env '["𝜉"] Int, Observables env '["β", "η", "γ", "ρ"] Double) => Int -> Popl -> Model env ts (Popl, [Popl])
+hmmSIRS :: (Lookups popl '["s", "i", "r"] Int,
+            Observables env '["𝜉"] Int, Observables env '["β", "η", "γ", "ρ"] Double)
+  -- | number of days
+  => Int
+  -- | initial population state
+  -> Record popl
+  -- | (final population state, intermediate population states)
+  -> Model env es (Record popl, [Record popl])
 hmmSIRS n = handleWriterM . hmmGen transPriorSIRS obsPriorSIR transSIRS obsSIR n
 
--- || (Section 3.2, Fig 4b) SIM from SIRS model: ([(s, i, r)], [𝜉])
+-- | Simulate from SIRS model: ([(s, i, r)], [𝜉])
 simSIRS :: Int -> Sampler ([(Int, Int, Int)], [Reported])
 simSIRS n_days = do
-  let sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #η := [0.05] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
-      sir_0      = Popl {s = 762, i = 1, r = 0}
+  -- Specify model input of 762 susceptible and 1 infected
+  let sir_0      = #s @= 762 <: #i @= 1 <: #r @= 0 <: emptyRecord
+  -- Specify model environment
+      sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #η := [0.05] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
+  -- Simulate an epidemic over n_days
   ((_, sir_trace), sim_env_out) <- SIM.simulate (hmmSIRS n_days sir_0) sim_env_in
+  -- Get the observed infections over n_days
   let 𝜉s :: [Reported] = get #𝜉 sim_env_out
-      sirs = map (\(Popl s i recov) -> (s, i, recov)) sir_trace
+  -- Get the true SIRS values over n_days
+      sirs = map (\sir -> (sir ^. s, sir ^. i, sir ^. r)) sir_trace
   return (sirs, 𝜉s)
 
--- || (Section 3.2) SIRSV (resusceptible + vacc) model
-data TransParamsSIRSV = TransParamsSIRSV {
-    betaP_SIRSV  :: Double, -- ^ Mean contact rate between susceptible and infected people
-    gammaP_SIRSV :: Double, -- ^ Mean recovery rate
-    etaP_SIRSV   :: Double, -- ^ Rate of resusceptible
-    omegaP_SIRSV :: Double  -- ^ Vaccination rate
-}
-
-data PoplV = PoplV {
-    s' :: Int,
-    i' :: Int,
-    r' :: Int,
-    v' :: Int
-} deriving Show
-
--- | SIRSV transition models
-transSI' :: TransModel env ts Double PoplV
-transSI' beta (PoplV s i r v) = do
-  let pop = s + i + r + v
-  dN_SI <- binomial' s (1 - exp ((-beta * fromIntegral i) / fromIntegral pop))
-  return $ PoplV (s - dN_SI) (i + dN_SI) r v
-
-transIR' :: TransModel env ts Double PoplV
-transIR' gamma (PoplV s i r v)  = do
-  dN_IR <- binomial' i (1 - exp (-gamma))
-  return $ PoplV s (i - dN_IR) (r + dN_IR) v
-
-transRS' :: TransModel env es Double PoplV
-transRS' eta (PoplV s i r v) = do
-  dN_RS <- binomial' r (1 - exp (-eta))
-  return $ PoplV (s + dN_RS) i (r - dN_RS) v
-
-transSV' :: TransModel env es Double PoplV
-transSV' omega (PoplV s i r v)  = do
-  dN_SV <- binomial' s (1 - exp (-omega))
-  return $  PoplV (s - dN_SV) i r (v + dN_SV )
-
-transSIRSV :: Member (Writer [PoplV]) ts => TransModel env ts TransParamsSIRSV PoplV
-transSIRSV (TransParamsSIRSV beta gamma omega eta) sirv = do
-  sirv' <- (transSI' beta  >=>
-            transIR' gamma >=>
-            transRS' eta   >=>
-            transSV' omega) sirv
-  tellM [sirv']
-  return sirv'
+{- | SIRSV model.
+-}
 
 -- | SIRSV transition prior
 transPriorSIRSV :: Observables env '["β", "γ", "ω", "η"] Double
-  => Model env ts TransParamsSIRSV
+  => Model env ts (Double, Double, Double, Double)
 transPriorSIRSV  = do
-  TransParamsSIRS pBeta pGamma pEta <- transPriorSIRS
+  (pBeta, pGamma, pEta) <- transPriorSIRS
   pOmega <- gamma 1 (1/16) #ω
-  return (TransParamsSIRSV pBeta pGamma pEta pOmega)
+  return (pBeta, pGamma, pEta, pOmega)
 
--- | SIRSV observation model
-obsSIRSV :: Observable env "𝜉" Int
-  => ObsModel env ts Double PoplV Reported
-obsSIRSV rho (PoplV _ i _ v)  = do
-  i <- poisson (rho * fromIntegral i) #𝜉
-  return i
+-- | Transition model from S to V
+transSV :: Lookups popl '["s", "v"] Int => TransModel env es Double (Record popl)
+transSV omega popl  = do
+  let (s_0, v_0) = (popl ^. s,  popl ^. v)
+  dN_SV <- binomial' s_0 (1 - exp (-omega))
+  return $ popl & s .~ (s_0 - dN_SV)
+                & v .~ (v_0 + dN_SV)
+
+-- | Transition model from S to I, I to R, R to S, and S to V
+transSIRSV :: Lookups popl '["s", "i", "r", "v"] Int => TransModel env ts (Double, Double, Double, Double) (Record popl)
+transSIRSV (beta, gamma, eta, omega) =
+  transSI beta >=> transIR gamma >=> transRS eta  >=> transSV omega
 
 -- | SIRSV as HMM
-hmmSIRSV ::  (Observables env '["𝜉"] Int, Observables env '["β", "γ", "η", "ω", "ρ"] Double) => Int -> PoplV -> Model env ts (PoplV, [PoplV])
-hmmSIRSV n = handleWriterM . hmmGen transPriorSIRSV obsPriorSIR transSIRSV obsSIRSV n
+hmmSIRSV :: (Lookups popl '["s", "i", "r", "v"] Int,
+             Observables env '["𝜉"] Int, Observables env '["β", "η", "γ", "ω", "ρ"] Double)
+  -- | number of days
+  => Int
+  -- | initial population state
+  -> Record popl
+  -- | (final population state, intermediate population states)
+  -> Model env es (Record popl, [Record popl])
+hmmSIRSV n_days = handleWriterM . hmmGen transPriorSIRSV obsPriorSIR transSIRSV obsSIR n_days
 
--- || (Section 3.2, Fig 4c) SIM from SIRSV model : ([(s, i, r, v)], [𝜉])
+-- | Simulate from SIRSV model : ([(s, i, r, v)], [𝜉])
 simSIRSV :: Int -> Sampler ([(Int, Int, Int, Int)], [Reported])
 simSIRSV n_days = do
-  let sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #η := [0.05] <:> #ω := [0.02] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
-      sirv_0      = PoplV {s' = 762, i' = 1, r' = 0, v' = 0}
-  ((_, sirv_trace), sim_env_out) <- SIM.simulate (hmmSIRSV n_days sirv_0) sim_env_in
+  -- Specify model input of 762 susceptible and 1 infected
+  let sir_0      = #s @= 762 <: #i @= 1 <: #r @= 0 <: #v @= 0 <: emptyRecord
+  -- Specify model environment
+      sim_env_in = #β := [0.7] <:> #γ := [0.009] <:> #η := [0.05] <:> #ω := [0.02] <:> #ρ := [0.3] <:> #𝜉 := [] <:> enil
+  -- Simulate an epidemic over n_days
+  ((_, sir_trace), sim_env_out) <- SIM.simulate (hmmSIRSV n_days sir_0) sim_env_in
+  -- Get the observed infections over n_days
   let 𝜉s :: [Reported] = get #𝜉 sim_env_out
-      sirvs = map (\(PoplV s i recov v) -> (s, i, recov, v)) sirv_trace
+  -- Get the true SIRSV values over n_days
+      sirvs = map (\sirv -> (sirv ^. s, sirv ^. i, sirv ^. r, sirv ^. v)) sir_trace
   return (sirvs, 𝜉s)